Client/Server/requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
psycopg2-binary>=2.9
django-environ>=0.9
whitenoise>=6.5
django-cors-headers>=4.0
gunicorn>=20.1
black==24.1.0
ruff==0.5.1
sentry-sdk>=1.8.0
django-csp>=3.0
boto3>=1.28
awscli>=1.29
django-storages>=1.14
django-extensions>=3.2
